backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Local AI with Ollama (No API keys needed!)
httpx==0.26.0  # For Ollama API calls
lancedb==0.20.0  # Local vector database (no C++ compilation needed!)
pyarrow>=14.0.0  # Required by lancedb

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
pypdf==4.0.1

# Data processing
pandas>=2.2.0
numpy>=2.0.0

# Environment and configuration
python-dotenv==1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Database and storage
motor>=3.6.0
pymongo>=4.10.0  # Compatible version with motor 3.6.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3

# Utilities
aiofiles==23.2.1
